+++ carp_mobile_sensing/lib/runtime/runtime.dart
@@ -15,6 +15,9 @@ import 'dart:convert';
 import 'dart:math';
 import 'dart:io';
 
+import 'package:ansicolor/ansicolor.dart';
+import 'dart:developer' as log;
+
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/foundation.dart';
 import 'package:json_annotation/json_annotation.dart';
@@ -74,20 +77,31 @@ class SensingException implements Exception {
   String toString() => '$runtimeType - $message';
 }
 
-/// Print a warning messages to the console.
+// AnsiPen pen =  AnsiPen()..white(bold: true);
+AnsiPen red = AnsiPen()..red();
+AnsiPen green = AnsiPen()..green();
+AnsiPen yellow = AnsiPen()..yellow();
+
+// print(pen("Bright white foreground") + " this text is default fg/bg");
+
+/// Add an information messages to the system log.
 void info(String message) =>
     (Settings().debugLevel.index >= DebugLevel.info.index)
-        ? print('\x1B[32m[CAMS INFO]\x1B[0m $message')
+        // ? print('\x1B[32m[CAMS INFO]\x1B[0m $message')
+        // ? print(green('[CAMS INFO] ') + message)
+        ? log.log('\x1B[32m[CAMS INFO]\x1B[0m $message')
         : 0;
 
-/// Pprint a warning messages to the console.
+/// Add a warning messages to the system log.
 void warning(String message) =>
     (Settings().debugLevel.index >= DebugLevel.warning.index)
-        ? print('\x1B[31m[CAMS WARNING]\x1B[0m  $message')
+        // ? print('\x1B[31m[CAMS WARNING]\x1B[0m  $message')
+        ? log.log('\x1B[31m[CAMS WARNING]\x1B[0m $message')
         : 0;
 
-/// Print a debug messages to the console.
+/// Add a debug messages to the system log.
 void debug(String message) =>
     (Settings().debugLevel.index >= DebugLevel.debug.index)
-        ? print('\x1B[33m[CAMS DEBUG]\x1B[0m $message')
+        // ? print('\x1B[33m[CAMS DEBUG]\x1B[0m $message')
+        ? log.log('\x1B[35m[CAMS DEBUG]\x1B[0m $message')
         : 0;